+++ NEWS.md
@@ -1,3 +1,5 @@
+# msSPChelpR (development version)
+
 # msSPChelpR 0.8.7
 
 ### New Features