+++ NEWS.md
@@ -2,6 +2,10 @@
 
 ### New Features
 * new function `calc_refrates()` to calculate age-, sex-, region-, year-specific reference rates from a long format dataframe with cancer cases that are counted for incident cases and then matched with a reference population. The resulting reference rates dataframe can directly be used with `sir_byfutime()` function.
+* functions gain new default `dattype = NULL` and thus are more flexible to take other source data types (Closes #73)
+
+### Breaking Changes
+* functions `asir`, `calc_futime*`, `calc_refrates`, `ir_crosstab_byfutime`, `pat_status*`, `renumber_time_id*`, and  `sir_byfutime` now by default are set to `dattype = NULL`. If you relied on automatic variable naming feature, you need to add `dattype = "seer"`or `dattype = "zfkd"` to your function call.
 
 ### Bug fixes
 * `sir_byfutime()`: