+++ NEWS.md
@@ -12,7 +12,7 @@
   * attributes with notes and problems are now correctly saved to `results_df`
 
 ### Internal
-* deprecated functions from `tidytable` package have been replaced (Closes #71)
+* deprecated functions from `tidytable` package have been replaced (Closes #71 and #74)
 
 
 # msSPChelpR 0.8.7